Small fixes

+++ gui/public/client/mail_delete.php
@@ -136,7 +136,7 @@ function client_deleteMail($mailId, $dmnProps)
 
 		iMSCP_Events_Manager::getInstance()->dispatch(iMSCP_Events::onBeforeDeleteMail, array('mailId' => $mailId));
 
-		if($cfg->PO_SERVER == 'dovecot') {
+		if(isset($cfg['PO_SERVER']) && $cfg['PO_SERVER'] == 'dovecot') {
 			$query = 'DELETE FROM `quota_dovecot` WHERE `username` = ?';
 			exec_query($query, $mailAddr);
 		}
